requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.52.0
httpx==0.27.2
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
aiofiles==23.2.1
python-jose==3.3.0
bcrypt==4.1.1
pydub==0.25.1
email-validator==2.1.0
aiohttp==3.9.1
websockets==11.0
